backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-jose[cryptography]
passlib[bcrypt] 
python-multipart
pydantic
pydantic-settings
email-validator
google-genai
python-dotenv
psycopg2-binary